requirements.txt
# Core Data Science
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
scipy>=1.9.0

# Deep Learning
torch>=1.12.0
pytorch-tabnet>=4.0
transformers>=4.21.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# XAI
shap>=0.41.0
lime>=0.2.0
captum>=0.6.0

# Web Framework
fastapi>=0.78.0
uvicorn>=0.18.0
pydantic>=1.9.0

# Database & Storage
sqlalchemy>=1.4.0
redis>=4.3.0
pymongo>=4.2.0

# Stream Processing
kafka-python>=2.0.0
pyspark>=3.3.0

# Monitoring & Logging
mlflow>=1.28.0
wandb>=0.13.0
loguru>=0.6.0

# Utilities
tqdm>=4.64.0
joblib>=1.1.0
pyyaml>=6.0
python-dotenv>=0.19.0
click>=8.1.0

# Development
jupyter>=1.0.0
notebook>=6.4.0
ipython>=8.4.0
pytest>=7.1.0
black>=22.6.0
flake8>=5.0.0

# Mobile App Dependencies
requests>=2.28.0
websockets>=10.3
aiofiles>=22.1.0